+++ src/openweather_sdk/__init__.py
@@ -291,6 +291,31 @@ class Client:
                 )
                 return forecast
 
+    def _get_location_forecast_daily_30_days(self, location):
+        lon, lat = self._get_location_coordinates(location)
+        return self._get_forecast_daily_30_days(lon, lat)
+
+    def _get_zip_code_forecast_daily_30_days(self, zip_code):
+        lon, lat = self._get_zip_code_coordinates(zip_code)
+        return self._get_forecast_daily_30_days(lon, lat)
+
+    def _get_forecast_daily_30_days(self, lon, lat):
+        if self.cache_size:
+            with self.lock:
+                if self.cache["forecast_30_days"]._is_relevant_info(lon, lat):
+                    return self.cache["forecast_30_days"]._get_info(lon, lat)
+
+        forecast_api = _ForecastAPI(lon=lon, lat=lat, appid=self.token)
+        forecast = forecast_api._get_forecast_daily_30_days()
+
+        if self.cache_size:
+            with self.lock:
+                self.cache["forecast_16_days"]._add_info(lon, lat, forecast)
+                logger.info(
+                    f"The client {self} has received data about 30 days forecast: {forecast}"
+                )
+                return forecast
+
     def get_location_weather(self, location, compact_mode=True):
         """
         Returns current weather in location (city name, state code (only for
@@ -385,7 +410,7 @@ class Client:
 
     def forecast_5_days(self, location=None, zip_code=None):
         """
-        Returns 5 day weather forecast data with 3-hour step at specified location.
+        Returns 5 days weather forecast data with 3-hour step at specified location.
         The location can be provided either as a combination of city name,
         state code (for the US), and country code separated by commas, or
         as a combination of zip/post code and country code separated by commas.
@@ -455,7 +480,7 @@ class Client:
 
     def forecast_daily_16_days(self, location=None, zip_code=None):
         """
-        Returns 16 day weather forecast data at specified location.
+        Returns 16 days weather forecast data at specified location.
         The location can be provided either as a combination of city name,
         state code (for the US), and country code separated by commas, or
         as a combination of zip/post code and country code separated by commas.
@@ -468,7 +493,7 @@ class Client:
             zip_code (str, optional): zip/post code and country code divided by comma.
         """
         logger.info(
-            f"The client {self} is being requested 5 day forecast in the location {location or zip_code}..."
+            f"The client {self} is being requested 16 days forecast in the location {location or zip_code}..."
         )
         if not self.is_alive:
             raise ClientDoesntExistException(self)
@@ -488,3 +513,39 @@ class Client:
                     "You need to specify zip code as a string"
                 )
             return self._get_zip_code_forecast_daily_16_days(zip_code)
+
+    def forecast_daily_30_days(self, location=None, zip_code=None):
+        """
+        Returns 30 days weather forecast data at specified location.
+        The location can be provided either as a combination of city name,
+        state code (for the US), and country code separated by commas, or
+        as a combination of zip/post code and country code separated by commas.
+        Please ensure the usage of ISO 3166 country codes.
+
+        Accessible with a "Developer" subscription and higher. See: https://openweathermap.org/full-price.
+
+        Args:
+            location (str, optional): city name, state code (only for the US) and country code divided by comma.
+            zip_code (str, optional): zip/post code and country code divided by comma.
+        """
+        logger.info(
+            f"The client {self} is being requested 30 days forecast in the location {location or zip_code}..."
+        )
+        if not self.is_alive:
+            raise ClientDoesntExistException(self)
+        if not location and not zip_code:
+            raise InvalidLocationException(
+                "You need to specify the location or postal code."
+            )
+        if location:
+            if not isinstance(location, str):
+                raise InvalidLocationException(
+                    "You need to specify the location as a string."
+                )
+            return self._get_location_forecast_daily_30_days(location)
+        if zip_code:
+            if not isinstance(zip_code, str):
+                raise InvalidLocationException(
+                    "You need to specify zip code as a string"
+                )
+            return self._get_zip_code_forecast_daily_30_days(zip_code)